server/requirements.txt
Flask==2.0.1
flask-mongoengine==1.0.0
grpcio==1.38.1
grpcio-tools==1.38.1
protobuf==3.17.3
Werkzeug==2.2.2